Merged PR 254: We should only have Wine and Other spirit types in Production Fermentation workflow
Fixing a bug where we should only have Wine and Other spirit types in Production Fermentation workflow.

The proper solution is to have filtering done on backend and this is the task tracking this work - Task 1846: Make sure that GetReportingSpiritTypes method does the filtering on backend, instead of front-end.

For sake of time saving, I am making the changes to be the same as it is implemented in Purchase Fermentation workflows

Related work items: #1835