Merged PR 265: Storage report would unexpectedly reduce Line 2 of Storage "Deposited in Bulk Storage" for GNS whenever it would get used for GIN
Storage report would unexpectedly reduce Line 2 of Storage "Deposited in Bulk Storage" for GNS whenever it would get used for GIN.
Another case where we would still need to have an old behavior is when we blend something within the same month. In that case, we would want to reduce Line 2 as it still happens in the same month.

Related work items: #1856