Merged PR 269: OnHandEndOfMonth in storage report is notbeing calculated properly
OnHandEndOfMonth in storage report is not being calculated properly.
The reason is that we would not allow adding up the proof if the record set from Purchased records contained a record with the same Purhase ID. As it turns out there cases where we do need to include those records with repeating Purchase IDs. For example, when those purchases were used in different time frames, I suspect. The fix is to remove dependency on Purchase and Production IDs lists and make linq queries return distinct results. This helped to only provide unique records in the records set.

Related work items: #1862