Merged PR 274: TaxWithdrawn rounding bug fix
The issue was that JavaScript was not properly rounding the difference between total available proof gallons and Tax Withdrawn.
In this specific case 42.17 - 27.94 = 19.179999999999996 which then would get truncated to 19.17. This change fixes it and the value is properly rounded to 19.18.
I have also done refactoring in ProcessingReport.cs file as part of this bug fix. The reason I did the refactoring was because this made my initial investigations easier.

No integration tests were added since the bug was on the front-end.

Related work items: #1866