Merged PR 281: fixing a bug in Processing Report that was not reproting correctly lines 31 and 47
The problem was in processing report class, method OnHandEndOfMonthPart2 where we did the calculations.
I simplified the method by removing redundant and erroneous query and replaced it with simple algebraic methods.
Also, made some changes around what we are computing on backend vs frontend of reporting. Task Task 1884: Remove computing of totals for part 2 of processing report with new members that were added as part of bug 1881

Related work items: #1881