Merged PR 325: bug fix
bug fix -
root cause - we are using onchange event in tax withdrawn control for calculating new proof gallon.

firing of onchange event requires for two things to be true. 1. the value in the control needs to change. 2. there has to be a focus change on the control.

In cases where I observed the bug to repro after I entered the value for tax withdrawn, I directly clicked on save button. This would send an update request to the backend before onchange event would get fired. In the other case, if I clicked anywhere in the edit window before I pressed save onchnage event would get fired thus updating to correct proof amount before the request is sent to the server.

Related work items: #2035